requirements.txt
requests
python-dotenv
black
streamlit